Fix deserialization errors hiding errors from other topics (#7077)
**User-Facing Changes**
None

**Description**

The IterablePlayer uses the `connectionId` as part of the key in the
problem manager. The source was not sending any `connectionId`, so
errors on multiple channels would override each other and only one was
visible at a time. With this change, each channel's latest error is
visible.


https://github.com/foxglove/studio/blob/3f87e3009acfb94020ad8270492225179ca6a5d3/packages/studio-base/src/players/IterablePlayer/IterablePlayer.ts#L652

Fixes FG-5631