Tolerate missing properties key in JSON schema (#7100)
**User-Facing Changes**

Not significant.

**Description**

This allows a simpler representation of a JSON schema representing "any
object".

When parsing a JSON schema, we assume the presence of `.properties`, but
this is not a required field. Currently, if I open an mcap file with a
JSON schema of `{"type":"object"}`, I see a confusing error. We could
produce a better error message. However, I'd like to document this
workaround for empty schemas, and accepting the above schema seems
preferable to requiring `{"type": "object", "properties": {}}`.