Avoid re-rendering player manager and message pipeline on global variable changes (#7154)
The PlayerManager and MessagePipeline bridge the react world with the
non-react player world. They sit relatively high up in the component
tree and re-rendering them is relatively "expensive".

Global variable changes don't actually impact the state of either
component, there's no need to re-render them and their children when
variables change. Instead we can directly update the player with the new
variables and avoid the re-render.