Restore state transition panel downsampling (#7205)
**User-Facing Changes**
Marginally speed up the rendering performance of the state transition
panel.

**Description**
This had broken at some point, certainly when I introduced
`useProvider`. Honestly though, downsampling doesn't do very much for
the state transition panel right now because distinct states are always
represented by at least one point; the maximum number of points is not
respected for some nuanced reasons. We will address this soon.

In the mean time though, this PR helps because when we zoom in, we will
(correctly) cull points outside of the viewport. We are not doing that
right now.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->